splitting melt file into melting and differential privacy portions, minor adjustment to figure generation, README update
