+++ README.md
@@ -2,11 +2,11 @@
 
 ## **1. Introduction**
 
-This is a library for processing scanned or digital EEO-1 and EEO-5 PDF reports as required by 141 of the Acts of 2024 (Massachusetts Salary Range Transparency Law).
+This is a library for processing scanned or digital EEO-1, EEO-4, and EEO-5 PDF reports as required by 141 of the Acts of 2024 (Massachusetts Salary Range Transparency Law).
 
 The repository also includes tools for post-processing, data aggregation and analysis of the extracted data.
 
-The repository provides custom parsing logic for different form types (e.g., EEO-1, EEO-5). This reduced parsing errors and improved extraction accuracy.
+The repository provides custom parsing logic for different form types (e.g., EEO-1, EEO-4, EEO-5). This reduced parsing errors and improved extraction accuracy.
 
 ---
 
@@ -14,27 +14,45 @@ The repository provides custom parsing logic for different form types (e.g., EEO
 
 <img src="ocr/architecture.png" alt="Pipeline Architecture" width="900"/>
 
-The pipeline processes EEO-1 and EEO-5 forms (PDF or images) in batches and includes the following stages:
+The pipeline processes EEO-1, EEO-4, and EEO-5 forms (PDF or images) in batches and includes the following stages:
 
-1.**Preprocessing** – Enhances image quality for improved OCR performance.  
-2.**Optical Character Recognition** – Extracts text using a deep learning-based OCR engine.  
-3.**Postprocessing & Parsing** – Segments content into structured fields.  
-4.**Validation & Cleaning** – Validates extracted data (e.g., zip codes, city names) against public datasets.  
-5.**Aggregation & Analysis** – Groups, aggregates, and exports data for reporting.
+1.**Preprocessing** – Classifies, deduplicates, and re-renders input files for improved OCR performance.  
+2.**Optical Character Recognition** – Extracts text and table contents using a deep learning-based OCR engine, outputting structured JSON.  
+3.**Postprocessing & Filtering** – Segments content into structured fields and filters records by form type and geography (Massachusetts).  
+4.**Validation & Handling** – Validates extracted data (e.g., ZIP codes, city names) and joins with NAICS industry labels, organizational size bins, and county-level identifiers to produce an enriched CSV.  
+5.**Melting** – Reshapes the wide-format CSV into a long format (one row per Race × Gender × dimension combination), materializing the full Cartesian product so all combinations are represented.  
+6.**Differential Privacy** – Applies OpenDP Laplace noise to the melted data, producing a differentially private main table and curated side tables.  
+7.**Figure Generation** – Consumes the differentially private tables to produce publication-ready charts and summary tables.
 
 
 ### **Pipeline Components**
 
-| Component           | Tool Used   | Description                                                    |
-|---------------------|-------------|----------------------------------------------------------------|
-| Image Preprocessing | OpenCV, PIL | Deduplication, formatting, scaling, padding                    |
-| OCR Engine          | DocTR       | Handwritten and printed text recognition                       |
-| Postprocessing      | Python      | Field extraction, form segmentation, validation and correction |
-| Data Aggregation    | Pandas      | CSV/JSON parsing, group-by, statistical summaries              |
+| Component           | Tool Used        | Description                                                          |
+|---------------------|------------------|----------------------------------------------------------------------|
+| Image Preprocessing | OpenCV, PIL      | Deduplication, formatting, scaling, padding                          |
+| OCR Engine          | DocTR            | Handwritten and printed text recognition                             |
+| Postprocessing      | Python           | Field extraction, form segmentation, validation and correction       |
+| Data Aggregation    | Pandas           | JSON parsing, NAICS/county joins, group-by, melt to long format      |
+| Differential Privacy| OpenDP           | Laplace noise on contingency tables with configurable epsilon budget  |
+| Figure Generation   | Matplotlib       | Bar charts, pie charts, and summary tables from DP output            |
 
 
 Each component is built as an independent, reusable module, facilitating extensibility and debugging.
 
+
+### **Per-Form-Type Scripts**
+
+The OCR postprocessing and data aggregation stages each have form-specific implementations. The table below shows coverage by form type.
+
+| Stage             | EEO-1                    | EEO-4                                          | EEO-5                              |
+|-------------------|--------------------------|------------------------------------------------|------------------------------------|
+| Filter            | `eeo1_filter.py`         | `eeo4_type1_filter.py`, `eeo4_type2_filter.py` | `eeo5_filter.py`                   |
+| Handler           | `eeo1_handler.py`        | `eeo4_handler.py`                              | `eeo5_handler.py`                  |
+| Melt              | `eeo1_melt.py`           | `eeo4_melt.py`                                 | `eeo5_melt.py`                     |
+| Differential Privacy | `eeo1_dp.py`          | `eeo4_dp.py`                                   | —                                  |
+| Figure Generation | `eeo1_figure_generation.py` | `eeo4_figure_generation.py`                 | —                                  |
+
+Filter scripts live under `ocr/postprocess/`. Handler, melt, and DP scripts live under `data_aggregation/`. Figure generation scripts live under `figure_generation/`.
 
 ---
 
@@ -84,7 +102,7 @@ For the purpose of this project, we explored 3 different OCR tools: Tesseract, N
 | [Donut](https://github.com/clovaai/donut)               | High accuracy on plaintext documents       | Yes                    | Yes                       | Requires GPU for reasonable inference time; needs to be fine-tuned for different tasks |
 | [DocTR](https://github.com/mindee/doctr)                | High accuracy across varied document types | Yes                    | No                        | No good support for handle handwritten forms                     |
 
-Since the EEO-1 and EEO-5 forms are structured and have a complex layout, and we only have CPU resources, we chose to use **DocTR** for the OCR engine.
+Since the EEO-1, EEO-4, and EEO-5 forms are structured and have a complex layout, and we only have CPU resources, we chose to use **DocTR** for the OCR engine.
 
 ---
 
@@ -95,16 +113,11 @@ Since the EEO-1 and EEO-5 forms are structured and have a complex layout, and we
 
 ---
 
-## **6. Ongoing Work**
-
-- Integrate a layout detection model for dynamic form segmentation  
-- Expand support for additional form types and formats  
-- Train a custom OCR model fine-tuned on EEO forms  
-- Build an interactive viewer for browsing OCR results
-
 ## Authors
+**Anthony Huang**: https://github.com/anthua20
+
 **Jida Li**: https://github.com/jidalii
 
-**Haodong Xu**: https://github.com/chuckhxu
-
 **Rohit Vemparala**: https://github.com/RVKarmani
+
+**Haodong Xu**: https://github.com/chuckhxu
