+++ figure_generation/eeo1_figure_config.py
