+++ figure_generation/eeo4_figure_config.py
