+++ figure_generation/figures/figure_15.py
@@ -3,7 +3,7 @@ import pandas as pd
 import matplotlib.pyplot as plt
 from matplotlib.patches import Wedge as MplWedge
 import numpy as np
-from eeo4_config import BAR_EDGE_COLOR, PIE_ORDER, RACE_COLORS
+from eeo4_figure_config import BAR_EDGE_COLOR, PIE_ORDER, RACE_COLORS
 
 def figure_15(jrg_df, output_dir, table_dir):
     EXPLODE = 0.06
