+++ figure_generation/figures/figure_27.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import matplotlib.pyplot as plt
-from eeo4_config import BAR_EDGE_COLOR, FEMALE_COLOR, MALE_COLOR
+from eeo4_figure_config import BAR_EDGE_COLOR, FEMALE_COLOR, MALE_COLOR
 
 def figure_27(wfrg_df, output_dir, table_dir):
