+++ figure_generation/figures/figure_5.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import matplotlib.pyplot as plt
 import textwrap
-from eeo1_config import BAR_EDGE_COLOR
+from eeo1_figure_config import BAR_EDGE_COLOR
 
 def figure_5(df, output_dir, table_dir):
     job_totals = df.groupby('JobCategory')['Count'].sum().sort_values(ascending=False)
