+++ figure_generation/figures/figure_6.py
@@ -3,7 +3,7 @@ import pandas as pd
 import matplotlib.pyplot as plt
 import numpy as np
 import textwrap
-from eeo1_config import BAR_EDGE_COLOR, FEMALE_COLOR, MALE_COLOR
+from eeo1_figure_config import BAR_EDGE_COLOR, FEMALE_COLOR, MALE_COLOR
 
 def figure_6(df, output_dir, table_dir):
     job_gender = df.groupby(['JobCategory', 'Gender'])['Count'].sum().unstack()
